Refactor CRUD moving hyrax/collections to hyrax/dashboard/collections.  Also adds a second show page under hyrax/dashcoard/collections that only those with edit access can view and use to perform management tasks for the collection.  The original show page under hyrax/collections has had management tasks removed in preparation of becoming the public branded landing page.
There are a few errors that were not fully resolved as part of this commit.  See...
