+++ app/controllers/hyrax/my_controller.rb
@@ -22,8 +22,7 @@ module Hyrax
     configure_facets
 
     before_action :authenticate_user!
-    before_action :enforce_show_permissions, only: :show
-    before_action :enforce_viewing_context_for_show_requests, only: :show
+    # before_action :enforce_show_permissions, only: :show
 
     # include the render_check_all view helper method
     helper Hyrax::BatchEditsHelper
