--- app/views/hyrax/collections/_button_create_collection.html.erb
@@ -1,1 +1,0 @@
-<%= button_to label, new_collection_path, method: :get, class: "btn btn-primary collection-add submits-batches", 'data-behavior'=>'hydra-collections', id: 'hydra-collection-add' %>
