+++ app/views/hyrax/dashboard/collections/_batch_edits_actions.html.erb
