+++ app/views/hyrax/dashboard/collections/_bookmark_control.html.erb
