+++ app/views/hyrax/dashboard/collections/_button_create_collection.html.erb
@@ -1,0 +1,1 @@
+<%= button_to label, hyrax.new_dashboard_collection_path, method: :get, class: "btn btn-primary collection-add submits-batches", 'data-behavior'=>'hydra-collections', id: 'hydra-collection-add' %>
