+++ app/views/hyrax/dashboard/collections/_button_remove_from_collection.html.erb
@@ -1,4 +1,4 @@
-<%= form_for collection, url: collection_path(collection), method: :put, as: 'collection' do |f| %>
+<%= form_for collection, url: hyrax.dashboard_collection_path(collection), method: :put, as: 'collection' do |f| %>
   <%= single_item_action_remove_form_fields(f, document) %>
   <%= f.submit label, class: "btn btn-primary collection-remove" %>
 <% end %>
