+++ app/views/hyrax/dashboard/collections/_default_group.html.erb
@@ -1,0 +1,24 @@
+<table class="table table-striped">
+  <caption class="sr-only"><%= t("hyrax.dashboard.my.sr.listing") %> <%= application_name %></caption>
+  <% unless params[:display_type] == 'list' %>
+    <thead>
+    <tr>
+      <th class="check-all"><label for="check_all" class="sr-only"><%= t("hyrax.dashboard.my.sr.check_all_label") %></label><%= render_check_all %></th>
+      <th><%= t("hyrax.dashboard.my.heading.title") %></th>
+      <th class="sorts-dash"><i id="<%= CatalogController.uploaded_field %>" class="<%= params[:sort] == "#{CatalogController.uploaded_field} desc" ? 'caret' : params[:sort] == "#{CatalogController.uploaded_field} asc" ? 'caret up' : ''%>"></i><%= t("hyrax.dashboard.my.heading.date_uploaded") %></th>
+      <th><%= t("hyrax.dashboard.my.heading.visibility") %></th>
+      <th><%= t("hyrax.dashboard.my.heading.action") %></th>
+    </tr>
+    </thead>
+  <% end %>
+  <tbody>
+  <% docs.each_with_index do |document, counter| %>
+    <% # Collection members may be collections or works %>
+    <% if document.collection? %>
+      <%= render 'list_collections', document: document, counter: counter %>
+    <% else %>
+      <%= render 'list_works', document: document, counter: counter, presenter: Hyrax::WorkShowPresenter.new(document, current_ability) %>
+    <% end %>
+  <% end %>
+  </tbody>
+</table>
