+++ app/views/hyrax/dashboard/collections/_document_list.html.erb
@@ -1,0 +1,12 @@
+<% @all_checked = false %>
+<% @user = current_user %>
+<% if documents.nil? || documents.empty? %>
+  <div class="alert alert-warning" role="alert"><%= t('hyrax.collection.document_list.no_visible_works') %></div>
+<% else %>
+  <% @document_list = documents %>
+  <% if params[:action] == "edit" %>
+    <%= render 'default_group', documents: documents, docs: documents %>
+  <% else %>
+    <%= render 'show_document_list', documents: documents %>
+  <% end %>
+<% end %>
