+++ app/views/hyrax/dashboard/collections/_edit_actions.html.erb
