+++ app/views/hyrax/dashboard/collections/_form.html.erb
@@ -12,7 +12,7 @@
       </li>
     <% end %>
   </ul>
-  <%= simple_form_for @form, html: { class: 'editor' } do |f| %>
+  <%= simple_form_for @form, url: [hyrax, :dashboard, @form], html: { class: 'editor' } do |f| %>
 
   <div class="tab-content">
       <div id="description" class="tab-pane active">
