+++ app/views/hyrax/dashboard/collections/_form_discovery.html.erb
