+++ app/views/hyrax/dashboard/collections/_list_collections.html.erb
