+++ app/views/hyrax/dashboard/collections/_list_works.html.erb
@@ -33,4 +33,8 @@
   <td class='text-center'>
     <span class="fa <%= current_user.trophies.where(work_id: document.id).exists? ? 'fa-star highlighted-work' : 'fa-star-o trophy-off' %>"></span></td>
   <td><%= render_visibility_link document %></td>
+
+  <td>
+    <%= render 'work_action_menu', document: document %>
+  </td>
 </tr>
