+++ app/views/hyrax/dashboard/collections/_show_actions.html.erb
@@ -2,7 +2,7 @@
 <div class="actions-controls-collections">
     <% if can? :edit, presenter.solr_document %>
       <%= link_to t('hyrax.collection.actions.edit.label'),
-                  edit_collection_path(presenter),
+                  edit_dashboard_collection_path(presenter),
                   title: t('hyrax.collection.actions.edit.desc'),
                   class: 'btn btn-default' %>
       <%= link_to t('hyrax.collection.actions.add_works.label'),
