+++ app/views/hyrax/dashboard/collections/_show_document_list.html.erb
@@ -1,0 +1,17 @@
+<table class="table table-striped">
+  <caption class="sr-only">List of items in this collection</caption>
+  <thead>
+  <tr>
+    <th>&nbsp;</th>
+    <th>Title</th>
+    <th>Date Uploaded</th>
+    <th>Visibility</th>
+    <% if current_user %>
+      <th>Action</th>
+    <% end %>
+  </tr>
+  </thead>
+  <tbody>
+    <%= render partial: 'show_document_list_row', collection: documents, as: :document %>
+  </tbody>
+</table>
