+++ app/views/hyrax/dashboard/collections/_show_document_list_menu.html.erb
@@ -1,0 +1,21 @@
+<div class="btn-group">
+  <button class="btn btn-default btn-sm dropdown-toggle" data-toggle="dropdown">Select <span class="caret"></span>
+  </button>
+  <ul class="dropdown-menu dropdown-menu-right">
+    <% if can? :edit, document %>
+      <li>
+        <%= link_to [main_app, :edit, document],
+                    class: "itemicon itemedit",
+                    title: t('hyrax.collection.document_list.edit'),
+                    id: "edit_work_link_#{document.id}" do %>
+          <i class='glyphicon glyphicon-pencil'></i> <%= t('hyrax.collection.document_list.edit') %>
+        <% end %>
+      </li>
+    <% end %>
+    <li>
+      <%= display_trophy_link(current_user, document.id) do |text| %>
+        <i class='glyphicon glyphicon-star'></i> <%= text %>
+      <% end %>
+    </li>
+  </ul>
+</div>
