+++ app/views/hyrax/dashboard/collections/_single_item_action_fields.html.erb
