+++ app/views/hyrax/dashboard/collections/_work_action_menu.html.erb
