+++ app/views/hyrax/dashboard/collections/edit_fields/_description.html.erb
