+++ app/views/hyrax/dashboard/collections/new.html.erb
@@ -6,6 +6,6 @@
 
 <div class="row">
   <div class="col-md-12">
-    <%= render 'hyrax/collections/form' %>
+    <%= render 'form' %>
   </div>
 </div>
