+++ app/views/hyrax/dashboard/collections/show.html.erb
@@ -2,16 +2,19 @@
 <div itemscope itemtype="http://schema.org/CollectionPage" class="row collection">
   <div class="col-sm-10 pull-right">
     <header>
-      <%= render 'collection_title', presenter: @presenter %>
+      <%= render 'hyrax/collections/collection_title', presenter: @presenter %>
     </header>
-    <%= render 'collection_description', presenter: @presenter %>
+    <%= render 'hyrax/collections/collection_description', presenter: @presenter %>
 
     <% unless has_collection_search_parameters? %>
       <%= render 'show_descriptions' %>
     <% end %>
   </div>
   <div class="col-sm-2">
-    <%= render 'media_display', presenter: @presenter %>
+    <%= render 'hyrax/collections/media_display', presenter: @presenter %>
+    <% unless has_collection_search_parameters? %>
+      <%= render 'show_actions', presenter: @presenter %>
+    <% end %>
   </div>
 </div>
 
@@ -23,11 +26,11 @@
         <%= t('.works_in_collection') %>
     <% end %>
   </h2>
-  <div class="col-xs-6 col-md-5 col-lg-6"><%= render 'search_form', presenter: @presenter %></div>
+  <div class="col-xs-6 col-md-5 col-lg-6"><%= render 'hyrax/collections/search_form', presenter: @presenter %></div>
 </div>
 
-<%= render 'sort_and_per_page', collection: @presenter %>
+<%= render 'hyrax/collections/sort_and_per_page', collection: @presenter %>
 
 <%= render_document_index @member_docs %>
 
-<%= render 'paginate' %>
+<%= render 'hyrax/collections/paginate' %>
