+++ app/views/hyrax/my/_collection_action_menu.html.erb
@@ -5,14 +5,14 @@
   </button>
   <ul role="menu" class="dropdown-menu dropdown-menu-right" aria-labelledby="dropdownMenu_<%= id %>">
     <li role="menuitem" tabindex="-1">
-      <%= link_to hyrax.edit_collection_path(id),
+      <%= link_to hyrax.edit_dashboard_collection_path(id),
                   class: 'itemicon itemedit',
                   title: t("hyrax.dashboard.my.action.edit_collection") do %>
         <i class="glyphicon glyphicon-pencil" aria-hidden="true"></i> <%= t("hyrax.dashboard.my.action.edit_collection") %>
       <% end %>
     </li>
     <li role="menuitem" tabindex="-1">
-      <%= link_to hyrax.collection_path(id),
+      <%= link_to hyrax.dashboard_collection_path(id),
                   class: 'itemicon itemtrash',
                   title: t("hyrax.dashboard.my.action.delete_collection"),
                   method: :delete,
