+++ app/views/hyrax/my/_search_header.html.erb
@@ -13,7 +13,7 @@
 </div>
 
 <div class="batch-info">
-  <%= render 'hyrax/collections/form_for_select_collection', user_collections: @user_collections %>
+  <%= render 'hyrax/dashboard/collections/form_for_select_collection', user_collections: @user_collections %>
 
   <div class="batch-toggle">
     <% session[:batch_edit_state] = "on" %>
