+++ app/views/hyrax/my/collections/index.html.erb
@@ -15,7 +15,7 @@
     <% if can?(:create, Collection) %>
       <%= link_to(
             t(:'helpers.action.collection.new'),
-            hyrax.new_collection_path,
+            hyrax.new_dashboard_collection_path,
             class: 'btn btn-primary'
           ) %>
     <% end %>
