+++ config/locales/hyrax.en.yml
@@ -398,27 +398,6 @@ en:
         title: "Add to collection"
         update: "Update Collection"
     collections:
-      edit:
-        header: "Edit Collection: %{title}"
-      form:
-        tabs:
-          description: "Description"
-          sharing:     "Sharing"
-          discovery:   "Discovery"
-      form_discovery:
-          open:
-            label:         "Open Access"
-            description:   "Visible to all"
-          institution:
-            label:         "Institution"
-            description:   "Restricted visibility to institution"
-          private:
-            label:         "Private"
-            description:   "Keep to myself with an option to share"
-          para1: "These settings determine who is able to discover and view this collection's landing page; they do not affect the visibility of items in the collection."
-          para2: "If you chose not to make this collection open access, you can still share the collection with specific users and groups using the Sharing tab."
-      new:
-        header: Create New Collection
       search_form:
         button_label: Go
         label: "Search Collection %{title}"
@@ -467,6 +446,30 @@ en:
       authorize_proxies:        "Authorize Proxies"
       breadcrumbs:
         admin:                  "Administration"
+      collections:
+        edit:
+          header: "Edit Collection: %{title}"
+        form:
+          tabs:
+            description: "Description"
+            sharing:     "Sharing"
+            discovery:   "Discovery"
+        form_discovery:
+          open:
+            label:         "Open Access"
+            description:   "Visible to all"
+          institution:
+            label:         "Institution"
+            description:   "Restricted visibility to institution"
+          private:
+            label:         "Private"
+            description:   "Keep to myself with an option to share"
+          para1: "These settings determine who is able to discover and view this collection's landing page; they do not affect the visibility of items in the collection."
+          para2: "If you chose not to make this collection open access, you can still share the collection with specific users and groups using the Sharing tab."
+        new:
+          header: Create New Collection
+        show:
+          works_in_collection: "Works in this Collection"
       create_work:              "Create Work"
       current_proxies:          "Current Proxies"
       heading_actions:
