+++ spec/views/hyrax/dashboard/collections/_form_discovery.erb_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe 'hyrax/collections/_form_discovery.html.erb', type: :view do
+RSpec.describe 'hyrax/dashboard/collections/_form_discovery.html.erb', type: :view do
   let(:collection) { Collection.new }
   let(:collection_form) { Hyrax::Forms::CollectionForm.new(collection, double, double) }
