+++ spec/views/hyrax/dashboard/collections/_show_descriptions.html.erb_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe 'hyrax/collections/_show_descriptions.html.erb', type: :view do
+RSpec.describe 'hyrax/dashboard/collections/_show_descriptions.html.erb', type: :view do
   context 'displaying a custom collection' do
     let(:collection_size) { 123_456_678 }
     let(:collection) do
