+++ spec/views/hyrax/dashboard/collections/_show_document_list.html.erb_spec.rb
@@ -1,13 +1,13 @@
-RSpec.describe 'hyrax/collections/_show_document_list.html.erb', type: :view do
+RSpec.describe 'hyrax/dashboard/collections/_show_document_list.html.erb', type: :view do
   let(:documents) { ["Hello", "World"] }
 
   before do
-    stub_template('hyrax/collections/_show_document_list_row.html.erb' => "<%= document %>")
+    stub_template('hyrax/dashboard/collections/_show_document_list_row.html.erb' => "<%= document %>")
   end
   context 'when not logged in' do
     it "renders the documents without an Action section" do
       allow(view).to receive(:current_user).and_return(nil)
-      render('hyrax/collections/show_document_list.html.erb', documents: documents)
+      render('show_document_list.html.erb', documents: documents)
       expect(rendered).to have_css('tbody', text: documents.join)
       expect(rendered).not_to have_css('th', text: 'Action')
     end
@@ -16,7 +16,7 @@ RSpec.describe 'hyrax/collections/_show_document_list.html.erb', type: :view do
   context 'when logged in' do
     it "renders the documents with an Action section" do
       allow(view).to receive(:current_user).and_return(true)
-      render('hyrax/collections/show_document_list.html.erb', documents: documents)
+      render('show_document_list.html.erb', documents: documents)
       expect(rendered).to have_css('tbody', text: documents.join)
       expect(rendered).to have_css('th', text: 'Action')
     end
