+++ spec/views/hyrax/dashboard/collections/show.html.erb_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe 'hyrax/collections/show.html.erb', type: :view do
+RSpec.describe 'hyrax/dashboard/collections/show.html.erb', type: :view do
   let(:document) do
     SolrDocument.new(id: 'xyz123z4',
                      'title_tesim' => ['Make Collections Great Again'],
@@ -16,17 +16,22 @@ RSpec.describe 'hyrax/collections/show.html.erb', type: :view do
     assign(:presenter, presenter)
 
     # Stub route because view specs don't handle engine routes
-    allow(view).to receive(:collection_path).and_return("/collection/123")
+    allow(view).to receive(:edit_dashboard_collection_path).and_return("/dashboard/collection/123/edit")
+    allow(view).to receive(:collection_path).and_return("/dashboard/collection/123")
 
-    stub_template '_search_form.html.erb' => 'search form'
+    stub_template 'hyrax/collections/_search_form.html.erb' => 'search form'
     stub_template 'hyrax/collections/_sort_and_per_page.html.erb' => 'sort and per page'
-    stub_template 'hyrax/collections/_document_list.html.erb' => 'document list'
+    stub_template '_document_list.html.erb' => 'document list'
     stub_template 'hyrax/collections/_paginate.html.erb' => 'paginate'
     stub_template 'hyrax/collections/_media_display.html.erb' => '<span class="fa fa-cubes collection-icon-search"></span>'
     render
   end
 
   it 'draws the page' do
+    expect(rendered).to have_selector 'h2', text: 'Actions'
+    expect(rendered).to have_link 'Edit'
+    expect(rendered).to have_link 'Delete'
+    expect(rendered).to have_link 'Add works'
     expect(rendered).to match '<span class="fa fa-cubes collection-icon-search"></span>'
   end
 end
